A Ruby Gosu Game
The history of the game is that the character is trapped in a nightmare
and to wake up from that nightmare he need to beat his main dream which
is about a zombie apocalypse, and the objective is to collect items for
turn on the car and get away, the necessary items are the keys, gas and
battery. The problem is that his nightmare has more layers and the items
are spread over them and the dreams (scenarios) can change anytime.
The game has two weapons, a pistol and a shotgun, wich you can alternate
picking up from the ground.